Moved ArrayIndexes to parser package.
